move qiniu mock to mock dir
